Late commit due to development rush. Ray tracing was already implemented, as well as the targets collision mechanic. The HUD is functional(already with bullet counter) and the easy mode of the game is already done. The score structure is already defined. The next steps are to define how to show the score information, create interaction buttons to control gameplay and create the other difficulties

+++ Teste_trabalho_CG_grauB/Vertex.h
@@ -3,9 +3,9 @@ class Vertex
 {
 public:
 
-	int pos_x;
-	int pos_y;
-	int pos_z;
+	float pos_x;
+	float pos_y;
+	float pos_z;
 
 	Vertex(float x, float y, float z)
 	{
